Add base class to test a foam tutorial
